Pom Enhancement to allow easy MyEclipse Development
MyEclipse uses maven <name> for standard maven project import name.  The
use of colon is not allowed in this process.  The alternative to a
normal import is to know to switch to use the artifactId.  This may or
may not be preferred.  Therefore, this PR is to simply change the colon
to a space with a dash.  It still retains the original intent and makes
myEclipse developers life a little easier when contributing to
richfaces.